Improved sync, added VSCL_BUMP cmd
Improved interprocess syncronization to ensure most recent value from
buffer is read.
Adjusted cameraProcess2 so that we transmit a TARGET bank angle to the
MAV. That target bank angle is taken to be the average bank angle during
the period of observation,  rounded to nearest 2 degrees.
Added VSCL_BUMP MAVlink command to allow on-the-fly adjustment of target
airspeed/altitude in FBWB mode. Untested.