Integrating bump command, UNSTABLE
Bump commands integrated, reads back commanded values from MAV.
Performance of camera processing is now unstable. I think the main
process (MAVProxy) is getting hung waiting forever to acquire the lock
from cameraProcess2 in tracking mode. The error seems to occur after
sending the "Update" command (add a print to check this next).
The other possibility is that the main process is crashing and I'm not
getting a crash message.